+++ web/src/ui_init.js
@@ -192,7 +192,11 @@ function initialize_compose_box() {
 }
 
 function initialize_message_feed_errors() {
-    $("#message_feed_errors_container").html(render_message_feed_errors());
+    $("#message_feed_errors_container").html(
+        render_message_feed_errors({
+            is_guest: current_user.is_guest,
+        }),
+    );
 }
 
 export function initialize_kitchen_sink_stuff() {